trash works sans <RestorePuzzle /> rendering
